+++ tests/helpers/testers.py
@@ -393,7 +393,7 @@ class MetricTester:
         metric_args: dict = {},
         **kwargs_update,
     ):
-        """Test if an metric can be used with half precision tensors on cpu
+        """Test if a metric can be used with half precision tensors on cpu
         Args:
             preds: torch tensor with predictions
             target: torch tensor with targets
@@ -416,7 +416,7 @@ class MetricTester:
         metric_args: dict = {},
         **kwargs_update,
     ):
-        """Test if an metric can be used with half precision tensors on gpu
+        """Test if a metric can be used with half precision tensors on gpu
         Args:
             preds: torch tensor with predictions
             target: torch tensor with targets
@@ -429,6 +429,39 @@ class MetricTester:
         _assert_half_support(
             metric_module(**metric_args), metric_functional, preds, target, device="cuda", **kwargs_update
         )
+
+    def run_differentiability_test(
+        self,
+        preds: torch.Tensor,
+        target: torch.Tensor,
+        metric_module: Metric,
+        metric_functional: Callable,
+        metric_args: dict = {},
+    ):
+        """Test if a metric is differentiable or not
+
+        Args:
+            preds: torch tensor with predictions
+            target: torch tensor with targets
+            metric_module: the metric module to test
+            metric_args: dict with additional arguments used for class initialization
+        """
+        # only floating point tensors can require grad
+        metric = metric_module(**metric_args)
+        if preds.is_floating_point():
+            preds.requires_grad = True
+            out = metric(preds[0], target[0])
+            assert metric.is_differentiable == out.requires_grad
+
+            if metric.is_differentiable:
+                # check for numerical correctness
+                assert torch.autograd.gradcheck(
+                    partial(metric_functional, **metric_args),
+                    (preds[0].double(), target[0])
+                )
+
+            # reset as else it will carry over to other tests
+            preds.requires_grad = False
 
 
 class DummyMetric(Metric):